Move contents of finesse.gui.hardware_set.device_connection into parent namespace
